feat: refactor request handling and remove unused code
Add new request classes for ticket types and categories. 
Remove unused authorization logic from controllers. 
Update event controller to include unauthenticated annotations. 
These changes improve code organization and maintainability.
