+++ app/Http/Requests/TicketType/StoreTicketTypeRequest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\TicketType;
 
 use Illuminate\Foundation\Http\FormRequest;
